backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database drivers
aiosqlite==0.19.0
asyncpg==0.29.0

# AI Providers
openai==1.12.0
anthropic==0.18.0
httpx==0.26.0
